Build alignment validation pipeline (Milestones B-H)
Major changes spanning the loader fix, the harness upgrade, native
output emission, the Taichi-side OUTPUT writer, the comparator, and the
matrix orchestrator:

  * mesh_loader_f1.py: thread a `dtype` argument through `load_hydro_mesh`
    and allocate XP/YP/SIDE/COSF/SINF/AREA/H_init/derived arrays in the
    caller-requested dtype. F1 fp32 entrypoint now passes np.float32 so
    geometry math runs in fp32 and matches native CUDA's
    `Real = float` `preCalculate`. Verified by
    scripts/diag/probe_loader_geometry.py: all geometry arrays now
    bit-match between F1 and F2 loaders in fp32 mode (was 116K of 207K
    cells diverging on AREA / SIDE / H_init).

  * F1/F2 fp32+fp64 step closures now accept an `on_step(step_index)`
    callback; default None preserves the existing all-at-once loop. The
    harness uses the callback to dump checkpoints from a single
    Taichi-process run.

  * scripts/check_correctness.py rewritten: --steps takes a comma list,
    runs Taichi once per case via on_step, runs native once per (case,
    step) in dump mode. Per-field stats include max/mean, p50/p90/p99/
    p99.9, threshold counts at 1e-7/1e-5/1e-3/1e-1, top-3 worst cells
    with KLAS context, plus W (=sqrt(U²+V²)). Conservation calculator
    emits mass, momentum X/Y, kinetic / potential energy, KLAS=10 and
    KLAS=1 inflow integrals. JSON reports land in results/alignment/
    and a Markdown SUMMARY.md aggregates them. fp64 / fp32-step1 /
    fp32-long thresholds are encoded in TOLERANCES; cancellation-heavy
    sums (momentum, BC inflow) get a looser bound than non-cancellation
    sums (mass, kinetic / potential energy).

  * cuda_native_impl/benchmark.cu and fp32_src/benchmark.cu: added
    --with-output and --ntoutput CLI flags. With --with-output the
    benchmark calls MeshData::outputToFile(jt, kt) at the configured
    NTOUTPUT cadence (defaulting to TIME.DAT or override via flag),
    producing the canonical OUTPUT/{H2U2V2,ZUV,SIDE,XY-TEC,TIMELOG}
    files used downstream for byte-level comparison.

  * taichi_impl/output_writer.py: Python implementation of native's
    `MeshData::outputToFile` format. Reproduces the 5-space indent,
    10-per-line / 100-per-block newline cadence, setw(10) fixed
    setprecision(4) numeric formatting, header line layout, FI angle
    formula, and SIDE.OUT one-time geometry dump. H2U2V2.OUT and
    ZUV.OUT use the cell-clamp rules from the native side (zero out
    H/U/V when H<=HM1, fall back to ZBC for Z).

  * scripts/compare_output_files.py: textual byte-equality diff plus
    H2/U2/V2/Z2/W2/FI block parser with per-(frame, label) max_abs /
    mean_abs / p99 / threshold counts. Self-comparison on the native
    OUTPUT directory verified to report 100% byte-identical.

  * scripts/run_alignment_full.sh: orchestrates the full case ×
    precision × step matrix (defaults to checkpoints
    1,100,900,7200,36000,72000) and aggregates per-run JSONs into
    SUMMARY.md.

Initial matrix run on checkpoints {1, 100, 900, 7200} (32 entries)
shows: 20 PASS / 12 FAIL.

  * fp64: F1_6.7K and F2_24K all pass at machine eps. F1/F2_207K pass
    through step=900 but show real divergence at step=7200 (H max
    1.3e-2, U max ~21) — boundary-day-rollover bug to investigate next.
  * fp32 step=1: all four cases pass (sub-ulp).
  * fp32 long-step: F2_24K passes through step=900 (hits a 1.78e-5
    kinetic-energy bound at step=7200 vs 1e-5 threshold). F1_6.7K
    breaks the momentum_y bound from step=100 (3.1e-2) because of
    cancellation in a near-symmetric flow. F1/F2_207K break klas1_inflow
    from step=100 (3e-3) and break U.p99 by step=7200 (1.95e-3).

The summary records these as discovered failures pending further
investigation. The pipeline itself (loader fix, harness upgrade,
output emission, comparator, orchestrator) is now in place for
subsequent rounds to drive.

results/alignment/*.json and SUMMARY.md are gitignored — they are
test outputs, regenerated by run_alignment_full.sh.

+++ scripts/diag/probe_loader_geometry.py
@@ -22,7 +22,9 @@ def main():
     import mesh_loader_f1 as mlf1
     import mesh_loader_f2 as mlf2
 
-    m1 = mlf1.load_hydro_mesh(mesh="20w")
+    # F1 loader now accepts a `dtype` kwarg (default fp64). Pass fp32 so the
+    # geometry math runs in fp32 and matches F2's fp32 output.
+    m1 = mlf1.load_hydro_mesh(mesh="20w", dtype=np.float32)
     m2 = mlf2.load_mesh(mesh="20w", dtype=np.float32)
 
     cel = m1["CEL"]
